requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.2.1
djangorestframework==3.12.2
gunicorn==20.0.4
numpy==1.20.1
pydantic==1.7.3
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
